==> 02_QC/03_norm_HVG_integration.Rmd <==
---
title: "Data normalization and pre-processing"
author: "Laura Jiménez Gracia"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: yes
    toc_float: yes
    number_sections: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE,
                      fig.align = "center")
```

# Introduction
Following with the results obtained in the previous notebook, we will now perform other pre-processing steps, such as data normalization, feature selection, scaling, dimensionality reduction, and data visualization prior to batch-effect correction (data integration). To do so, we will follow the current best practices for scRNA-seq QC described in [Luecken al. Mol Syst Biol (2018)](https://doi.org/10.15252/msb.20188746) and adapt some workflows from [Satija Lab](https://satijalab.org/seurat/vignettes.html).

## Objective
In this Rmarkdown document, we are going to perform the previous pre-processing steps.

# Pre-processing
## Libraries
```{r warning = FALSE, message = FALSE}
library(tidyverse)
library(Seurat)
library(gt)
library(harmony)
```

## Parameters
Here we will define and load objects and functions that will be used throughout the document.
```{r}
# Paths
path_project_metadata <- here::here("01_cellranger_mapping/data/VEIGAEST_metadata.csv")
path_r_objects <- here::here("02_QC/results/R_objects")

# Functions
source(here::here("bin/utils.R"))

# Parameters
confounder_variables <- c("treatment")
confounder_names <- c("Treatment")

qc_metrics <-  c("nCount_RNA", "nFeature_RNA", "S.Score", "G2M.Score", "pct_mt", "pct_rb")

# Color palette
color_palette <- Polychrome::createPalette(25, c("#fc6060", "#74f774", "#7c7cfc"))
names(color_palette) <- NULL
#Polychrome::swatch(color_palette)
```

## Load data
The data used in this Rmarkdown document comes from **VEIGAEST** dataset.
```{r}
# Load all metadata information
metadata <- read.csv(path_project_metadata)

metadata <- metadata %>% filter(type == "cDNA")

print("Libraries metadata")
DT::datatable(metadata, options = list(scrollX = TRUE))
```


```{r}
# Load Seurat object
seurat_obj <- readRDS(paste0(path_r_objects, "/VEIGAEST_alldata_filtered.rds"))
seurat_obj
```

## Explore QC results
Here, we show an overview of the scRNA-seq data obtained after the quality control.
```{r}
table_qc_gex(seurat_obj@meta.data, subtitle = "After cell QC filtering out")
```

```{r fig.height=5, fig.width=16}
gg_qc_by_gemid <- VlnPlot(seurat_obj,
                          features = c("nCount_RNA", "nFeature_RNA", "pct_mt", "scrublet_doublet_scores"),
                          group.by = "treatment",
                          pt.size = 0,
                          ncol = 4,
                          log = TRUE
                          )

gg_qc_by_gemid
```

# Data normalization
Count depths for identical cells can be different due to the variability inherent in the capture, RT and sequencing of a mRNA molecule. Thus, to avoid detecting gene expression differences based on count data when comparing cells we will apply a global-scaling normalization by sequencing depth. 

This method divides the feature expression measurements (each count) by the library size of the cell (total number of UMIs), multiplies this by a scale factor (10,000) and log-transforms it.

```{r}
# Normalization
seurat_obj <- NormalizeData(
  seurat_obj,
  normalization.method = "LogNormalize",
  scale.factor = 10000)

seurat_obj[["RNA"]]@data[1:10, 1:30]
```

## Computing cell cycle score
To assess cell cycle effects in the next steps, it is important to compute the cell cycle score once the data is normalized. To do so, we will assign each cell a score based on its expression of G2/M and S phase markers. This value will be stored in the Seurat metadata, and we will analyze it after the dimensionality reduction step is performed.
```{r}
mmus_s = gprofiler2::gorth(cc.genes.updated.2019$s.genes, source_organism = "hsapiens", target_organism = "mmusculus")$ortholog_name
mmus_g2m = gprofiler2::gorth(cc.genes.updated.2019$g2m.genes, source_organism = "hsapiens", target_organism = "mmusculus")$ortholog_name

# Testing for cell_cycle effects
seurat_obj <- CellCycleScoring(
  seurat_obj,
  s.features = mmus_s,
  g2m.features = mmus_g2m)
```


# Feature selection and data scaling
Next, we will filter a subset of features that exhibit high cell-to-cell variation to keep only highly variable genes that are “informative” of the variability in the dataset, which will be further used in downstream analysis.

```{r}
# Summary statistics of the number of features (genes) by cell
summary(Matrix::colSums(seurat_obj@assays$RNA@data[, ] > 0))
```

```{r fig.width=7, fig.height=5}
# Identify highly variable features
seurat_obj <- FindVariableFeatures(
  seurat_obj,
  nfeatures = 3000)

# Identify the 20 most highly variable genes
top20_hvf <- head(VariableFeatures(seurat_obj), 20)
top20_hvf

# Plot variable features
LabelPoints(VariableFeaturePlot(seurat_obj),
            points = top20_hvf,
            repel = TRUE)
```

Moreover, we will also apply a linear transformation `scaling` prior to dimensional reduction to avoid highly-expressed genes to dominate in the analysis. The scaling step shifts the expression of each gene, so that the mean expression across cells is 0 and the variance is 1.
```{r}
# Scaling
seurat_obj <- ScaleData(seurat_obj)
```

# Dimensionality reduction
Next, we will perform a linear dimensionality reduction using the `PCA` algorithm to reduce the high-dimension of data (we have many genes, and some of them are redundant).
```{r fig.width=7, fig.height=5}
# Run PCA
seurat_obj <- RunPCA(
  seurat_obj,
  ndims.print = 1:6,
  nfeatures.print = 10)

# Visualizing PCA results
# VizDimLoadings(seurat_obj, dims = 1:2, reduction = "pca")
```

After applying PCA, we will ensure that the first PCs capture cell-to-cell variability and not differences associated to batch-effects or by cell cycle phase.
```{r fig.width=7, fig.height=5}
gg_pca_by_lib <- DimPlot(
  seurat_obj,
  reduction = "pca",
  group.by = "gem_id") +
  labs(title = "", color = "Library")

gg_pca_by_lib
```

```{r fig.width=12, fig.height=5}
# Visualize the PCA reduction, grouping by cell cycle phase
gg_pca_by_cellcycle <- DimPlot(
  seurat_obj,
  reduction = "pca",
  group.by = "Phase",
  split.by = "Phase",
  pt.size = 0.01) + NoLegend()

gg_pca_by_cellcycle
```
Based on the previous plots, we do not observe significant batch-effects but specific cell cycle phase effects. Thus, we might regress out the variation due to cell cycle.

At this point, we will determine how many PCs should be considered for data visualization in 2 dimensions and for the downstream cluster analysis. It important to ensure that we are capturing the majority of the variation (or cell types) present in our dataset.

```{r fig.width=7, fig.height=5}
# Elbowplot to select relevant PCs to analyse the dataset
ElbowPlot(seurat_obj, ndims = 50, reduction = "pca")
```

In this particular dataset, we will use the first 30 PCs to visualize data and perform the clustering.

# Data visualization
For visualization purposes, it is standard practice to use the Uniform Approximation and Projection method (UMAP), which is a non-linear dimensionality reduction method that reduce the data dimensionality into a 2-dimension space.

```{r}
# Run UMAP
seurat_obj <- seurat_obj %>% 
  RunUMAP(reduction = "pca", dims = 1:30)
```

## Assessing confounder variables

```{r fig.width=7, fig.height=5}
# Visualizing UMAP
gg_umap_uncorrected <- purrr::map2(confounder_variables, confounder_names,
                                          function(confounder, title) {
  p <- DimPlot(seurat_obj, 
               reduction = "umap",
               group.by = confounder,
               pt.size = 0.1) +
    labs(title=title) +
    scale_color_manual(values = color_palette)
  p
})

cowplot::plot_grid(plotlist = gg_umap_uncorrected, ncol = 1)
```

Here, from the UMAP mapping, we can see the presence of library/donor-specific effects that should not be present. Thus, we will integrate different libraries to overcome the batch effect.

## Assessing cell cycle effects
```{r fig.width=7, fig.height=5}
gg_umap_by_phase <- DimPlot(
  seurat_obj,
  reduction = "umap",
  group.by = "Phase",
  pt.size = 0.01) +
  labs(color = "Cell Cycle Phase")

gg_umap_by_phase
```
Finally, we ensure that there is no batch effect associated to the cell cycle phase.


## Assessing sources of uninterested variation
Next we will explore additional metrics, such as the number of UMIs and genes per cell, S-phase and G2M-phase markers, and mitochondrial / ribosomal gene expression on the UMAP embedding. Finally, we will also check the scrublet doublet score.
```{r fig.width=14, fig.height=15}
# Determine metrics to plot present in seurat_integrated@meta.data
gg_qc_umap <- FeaturePlot(
  seurat_obj,
  reduction = "umap",
  features = qc_metrics,
  order = TRUE,
  ncol = 2)

gg_qc_umap
```


```{r fig.width=14, fig.height=5}
gg_doubletscore_umap <- FeaturePlot(
  seurat_obj,
  reduction = "umap",
  features = "scrublet_doublet_scores",
  order = TRUE)

gg_doubletpred_umap <- DimPlot(
  seurat_obj,
  reduction = "umap",
  group.by = "scrublet_predicted_doublet")

cowplot::plot_grid(gg_doubletscore_umap,
                   gg_doubletpred_umap, ncol = 2)
```

## Assessing gene markers expression
```{r fig.width=30, fig.height=20}
# Immune fraction
FeaturePlot(seurat_obj,
            features = c("Ptprc", # CD45+, immune cells
             "Cd3d", # T cells
             "Cd4", "Foxp3", "Sell", "Irf7", "Klrg1", # CD4+ T cells, Treg
             "Cd8a", "Nkg7", # CD8+ T cells, cytotoxic
             "Klrd1", "Il7r", # NK cells
             "Cd79a", # B cells
             "Ms4a1", "Jchain", # Naive vs Plasma cells
             "Lyz2", "Cd14", "Fcgr3", "Plac8", # Mono/macro
             "Clec9a", "Clec10a", "Ftl1", "Gzmb", # DCs
             "S100a9", # Neutrophils
             "Cd200r3", # Mast cells
             "H2-Ab1" # MHC-antigen presenting
             ),
            order = T,
            ncol = 5)
```

# Save processed Seurat objects
```{r}
# Save processed lognorm Seurat object
saveRDS(seurat_obj, paste0(path_r_objects, "/VEIGAEST_alldata_lognorm_processed.rds"))
```

# Integration with harmony
The integration approach was carried out with [Harmony](https://www.nature.com/articles/s41592-019-0619-0). [Chazarra-Gil *et al*](https://www.biorxiv.org/content/10.1101/2020.05.22.111211v2), which was reported to be among the best scRNA-seq integration methods, especially in terms of scalability, in a recent integration benchmarking study.

```{r}
seurat_obj_integrated <- seurat_obj %>%
  RunHarmony(group.by.vars = "gem_id")

# Elbowplot to select relevant PCs to analyse the dataset
ElbowPlot(seurat_obj_integrated, ndims = 50, reduction = "harmony")
```

As seen in the previous notebook, we will use the first 30 PCs to visualize data and perform the clustering.
```{r}
# Dimensionality reduction
seurat_obj_integrated <- seurat_obj_integrated %>%
  RunUMAP(reduction = "harmony", dims = 1:30)
```


## Assessing confounder variables
```{r fig.width=7, fig.height=5}
# Visualizing UMAP
gg_umap_corrected <- purrr::map2(confounder_variables, confounder_names,
                                          function(confounder, title) {
  p <- DimPlot(seurat_obj_integrated, 
               reduction = "umap",
               group.by = confounder,
               pt.size = 0.01) +
    labs(title=title) +
    scale_color_manual(values=color_palette)
  p
})

cowplot::plot_grid(plotlist = gg_umap_corrected, ncol = 1)
```

Here, from the UMAP mapping, we cannot see the presence of library-specific effects that should not be present. But, we clearly see treatment/disease differences. Thus, we will integrate the different samples to harmonize these cells.

## Assessing cell cycle effects
```{r fig.width=7, fig.height=5}
gg_umap_by_phase <- DimPlot(
  seurat_obj_integrated,
  reduction = "umap",
  group.by = "Phase",
  pt.size = 0.01) +
  labs(color = "Cell Cycle Phase")

gg_umap_by_phase
```
Finally, we ensure that there is no batch effect associated to the cell cycle phase.


## Assessing sources of uninterested variation
Next we will explore additional metrics, such as the number of UMIs and genes per cell, S-phase and G2M-phase markers, and mitochondrial / ribosomal gene expression on the UMAP embedding.

```{r fig.width=14, fig.height=15}
gg_qc_umap <- FeaturePlot(
  seurat_obj_integrated,
  reduction = "umap",
  features = qc_metrics,
  order = TRUE,
  ncol = 2)

gg_qc_umap
```


Finally, we'll also check for the doublet score and doublet prediction obtained using `scrublet`.
```{r fig.width=14, fig.height=5}
gg_doubletscore_umap <- FeaturePlot(
  seurat_obj_integrated,
  reduction = "umap",
  features = "scrublet_doublet_scores",
    cols = c("lightgray", "darkblue"),
  order = TRUE)

gg_doubletpred_umap <- DimPlot(
  seurat_obj_integrated,
  reduction = "umap",
  group.by = "scrublet_predicted_doublet")

cowplot::plot_grid(gg_doubletscore_umap,
                   gg_doubletpred_umap, ncol = 2)
```

## Assessing gene markers expression
```{r fig.width=30, fig.height=20}
# Immune fraction
FeaturePlot(seurat_obj_integrated,
            features = c("Ptprc", # CD45+, immune cells
             "Cd3d", # T cells
             "Cd4", "Foxp3", "Sell", "Irf7", "Klrg1", # CD4+ T cells, Treg
             "Cd8a", "Nkg7", # CD8+ T cells, cytotoxic
             "Klrd1", "Il7r", # NK cells
             "Cd79a", # B cells
             "Ms4a1", "Jchain", # Naive vs Plasma cells
             "Lyz2", "Cd14", "Fcgr3", "Plac8", # Mono/macro
             "Clec9a", "Clec10a", "Ftl1", "Gzmb", # DCs
             "S100a9", # Neutrophils
             "Cd200r3", # Mast cells
             "H2-Ab1" # MHC-antigen presenting
             ),
            order = T,
            ncol = 5)
```

<!-- # Getting data to evaluate integration -->
<!-- ```{r} -->
<!-- # Save confounders metadata and PCA coordinates to evaluate integration -->
<!-- confounders_df <- seurat_obj@meta.data[, confounder_variables] -->
<!-- saveRDS(confounders_df, file = paste0(path_r_objects, "/VEIGAEST_confounders.rds")) -->

<!-- # Get embeddings from non-integrated data -->
<!-- saveRDS(seurat_obj@reductions$pca@cell.embeddings[,1:20], -->
<!--         file = paste0(path_r_objects, "/VEIGAEST_pca_uncorrected.rds")) -->

<!-- # Get embeddings from integrated data -->
<!-- saveRDS(seurat_obj_integrated@reductions$harmony@cell.embeddings[,1:20], -->
<!--         file = paste0(path_r_objects, "/VEIGAEST_pca_integrated_harmony.rds")) -->
<!-- ``` -->

# Save processed Seurat objects
```{r}
# Save processed lognorm Seurat object
saveRDS(seurat_obj_integrated, paste0(path_r_objects, "/VEIGAEST_harmony_processed.rds"))
#seurat_obj_integrated <- readRDS(paste0(path_r_objects, "/VEIGAEST_harmony_processed.rds"))
```

# Session Info
```{r}
sessionInfo()
```
